feature/schema_company: add schema, crud, endpoint for Company
