--- alembic/versions/01_tabitadminuser_licensetype.py
@@ -1,59 +1,0 @@
-"""TabitAdminUser, LicenseType
-
-Revision ID: 01
-Revises:
-Create Date: 2025-01-25 02:10:39.764740
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import fastapi_users_db_sqlalchemy
-
-
-# revision identifiers, used by Alembic.
-revision: str = '01'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('licensetype',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=100), nullable=False),
-    sa.Column('license_term', sa.Interval(), nullable=False),
-    sa.Column('max_admins_count', sa.Integer(), nullable=False),
-    sa.Column('max_employees_count', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('tabitadminuser',
-    sa.Column('name', sa.String(length=100), nullable=False),
-    sa.Column('surname', sa.String(length=100), nullable=False),
-    sa.Column('patronymic', sa.String(length=100), nullable=False),
-    sa.Column('phone_number', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('id', fastapi_users_db_sqlalchemy.generics.GUID(), nullable=False),
-    sa.Column('email', sa.String(length=320), nullable=False),
-    sa.Column('hashed_password', sa.String(length=1024), nullable=False),
-    sa.Column('is_active', sa.Boolean(), nullable=False),
-    sa.Column('is_superuser', sa.Boolean(), nullable=False),
-    sa.Column('is_verified', sa.Boolean(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_tabitadminuser_email'), 'tabitadminuser', ['email'], unique=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_tabitadminuser_email'), table_name='tabitadminuser')
-    op.drop_table('tabitadminuser')
-    op.drop_table('licensetype')
-    # ### end Alembic commands ###
