--- alembic/versions/03_company.py
@@ -1,67 +1,0 @@
-"""Company
-
-Revision ID: 03
-Revises: 02
-Create Date: 2025-01-25 02:18:19.895821
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import fastapi_users_db_sqlalchemy
-
-
-# revision identifiers, used by Alembic.
-revision: str = '03'
-down_revision: Union[str, None] = '02'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('company',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('logo', sa.String(length=2048), nullable=False),
-    sa.Column('license_id', sa.Integer(), nullable=True),
-    sa.Column('max_admins_count', sa.Integer(), nullable=False),
-    sa.Column('max_employees_count', sa.Integer(), nullable=False),
-    sa.Column('start_license_time', sa.DateTime(), nullable=True),
-    sa.Column('end_license_time', sa.DateTime(), nullable=True),
-    sa.Column('is_active', sa.Boolean(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['license_id'], ['licensetype.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('department',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.Column('company_id', sa.Integer(), nullable=False),
-    sa.Column('supervisor_id', fastapi_users_db_sqlalchemy.generics.GUID(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['company_id'], ['company.id'], ),
-    sa.ForeignKeyConstraint(['supervisor_id'], ['usertabit.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('company_id', 'name', name='uq_company_department_name'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_unique_constraint(None, 'taguser', ['id'])
-    op.add_column('usertabit', sa.Column('company_id', sa.Integer(), nullable=False))
-    op.create_foreign_key(None, 'usertabit', 'company', ['company_id'], ['id'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'usertabit', type_='foreignkey')
-    op.drop_column('usertabit', 'company_id')
-    op.drop_constraint(None, 'taguser', type_='unique')
-    op.drop_table('department')
-    op.drop_table('company')
-    # ### end Alembic commands ###
