--- alembic/versions/04_department.py
@@ -1,40 +1,0 @@
-"""Department
-
-Revision ID: 04
-Revises: 03
-Create Date: 2025-01-25 02:19:44.356755
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '04'
-down_revision: Union[str, None] = '03'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_unique_constraint(None, 'company', ['id'])
-    op.create_unique_constraint(None, 'department', ['id'])
-    op.add_column('usertabit', sa.Column('current_department_id', sa.Integer(), nullable=False))
-    op.add_column('usertabit', sa.Column('last_department_id', sa.Integer(), nullable=True))
-    op.create_foreign_key(None, 'usertabit', 'department', ['current_department_id'], ['id'])
-    op.create_foreign_key(None, 'usertabit', 'department', ['last_department_id'], ['id'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'usertabit', type_='foreignkey')
-    op.drop_constraint(None, 'usertabit', type_='foreignkey')
-    op.drop_column('usertabit', 'last_department_id')
-    op.drop_column('usertabit', 'current_department_id')
-    op.drop_constraint(None, 'department', type_='unique')
-    op.drop_constraint(None, 'company', type_='unique')
-    # ### end Alembic commands ###
