--- alembic/versions/05_problem_meeting_task_and_etс.py
@@ -1,225 +1,0 @@
-"""Problem, Meeting, Task and etс
-
-Revision ID: 05
-Revises: 04
-Create Date: 2025-01-25 02:39:40.742094
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import fastapi_users_db_sqlalchemy
-
-# revision identifiers, used by Alembic.
-revision: str = '05'
-down_revision: Union[str, None] = '04'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('problem',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('color', sa.Enum('RED', 'ORANGE', 'YELLOW', 'GREEN', 'BLUE', 'DARK_BLUE', 'VIOLET', 'BROWN', 'GRAY', 'BLACK', 'WHITE', 'PINK', 'BEIGE', 'VINOUS', 'PURPLE', name='colorproblem'), nullable=False),
-    sa.Column('type', sa.Enum('A', 'B', 'C', 'D', 'E', 'F', 'G', name='typeproblem'), nullable=False),
-    sa.Column('status', sa.Enum('NEW', 'IN_PROGRESS', 'SUSPENDED', 'COMPLETED', name='statusproblem'), nullable=False),
-    sa.Column('owner_id', fastapi_users_db_sqlalchemy.generics.GUID(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['owner_id'], ['usertabit.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('associationuserproblem',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('left_id', fastapi_users_db_sqlalchemy.generics.GUID(), nullable=False),
-    sa.Column('right_id', sa.Integer(), nullable=False),
-    sa.Column('status', sa.Boolean(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['left_id'], ['usertabit.id'], ),
-    sa.ForeignKeyConstraint(['right_id'], ['problem.id'], ),
-    sa.PrimaryKeyConstraint('id', 'left_id', 'right_id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('fileproblem',
-    sa.Column('problem_id', sa.Integer(), nullable=False),
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('file_path', sa.String(length=2048), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['problem_id'], ['problem.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('meeting',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('title', sa.String(length=255), nullable=False),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('problem_id', sa.Integer(), nullable=False),
-    sa.Column('owner_id', fastapi_users_db_sqlalchemy.generics.GUID(), nullable=True),
-    sa.Column('date_meeting', sa.Date(), nullable=False),
-    sa.Column('status', sa.Enum('NEW', 'NOT_HELD', 'HELD', 'SUSPENDED', name='statusmeeting'), nullable=False),
-    sa.Column('place', sa.String(length=255), nullable=False),
-    sa.Column('transger_counter', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['owner_id'], ['usertabit.id'], ),
-    sa.ForeignKeyConstraint(['problem_id'], ['problem.id'], ),
-    sa.PrimaryKeyConstraint('id', 'problem_id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('messagefeed',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('problem_id', sa.Integer(), nullable=False),
-    sa.Column('owner_id', fastapi_users_db_sqlalchemy.generics.GUID(), nullable=True),
-    sa.Column('text', sa.String(), nullable=False),
-    sa.Column('important', sa.Boolean(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['owner_id'], ['usertabit.id'], ),
-    sa.ForeignKeyConstraint(['problem_id'], ['problem.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('task',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('date_completion', sa.Date(), nullable=False),
-    sa.Column('owner_id', fastapi_users_db_sqlalchemy.generics.GUID(), nullable=True),
-    sa.Column('problem_id', sa.Integer(), nullable=False),
-    sa.Column('status', sa.Enum('NEW', 'IN_PROGRESS', 'NOT_ACCEPTED', 'COMPLETED', name='statustask'), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['owner_id'], ['usertabit.id'], ),
-    sa.ForeignKeyConstraint(['problem_id'], ['problem.id'], ),
-    sa.PrimaryKeyConstraint('id', 'problem_id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('associationusermeeting',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('left_id', fastapi_users_db_sqlalchemy.generics.GUID(), nullable=False),
-    sa.Column('right_id', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['left_id'], ['usertabit.id'], ),
-    sa.ForeignKeyConstraint(['right_id'], ['meeting.id'], ),
-    sa.PrimaryKeyConstraint('id', 'left_id', 'right_id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('associationusertask',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('left_id', fastapi_users_db_sqlalchemy.generics.GUID(), nullable=False),
-    sa.Column('right_id', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['left_id'], ['usertabit.id'], ),
-    sa.ForeignKeyConstraint(['right_id'], ['task.id'], ),
-    sa.PrimaryKeyConstraint('id', 'left_id', 'right_id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('commentfeed',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('message_id', sa.Integer(), nullable=False),
-    sa.Column('owner_id', fastapi_users_db_sqlalchemy.generics.GUID(), nullable=True),
-    sa.Column('text', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['message_id'], ['messagefeed.id'], ),
-    sa.ForeignKeyConstraint(['owner_id'], ['usertabit.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('filemeeting',
-    sa.Column('meeting_id', sa.Integer(), nullable=False),
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('file_path', sa.String(length=2048), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['meeting_id'], ['meeting.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('filemessage',
-    sa.Column('message_id', sa.Integer(), nullable=False),
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('file_path', sa.String(length=2048), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['message_id'], ['messagefeed.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('filetask',
-    sa.Column('task_id', sa.Integer(), nullable=False),
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('file_path', sa.String(length=2048), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['task_id'], ['task.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('resultmeeting',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('meeting_id', sa.Integer(), nullable=False),
-    sa.Column('owner_id', fastapi_users_db_sqlalchemy.generics.GUID(), nullable=True),
-    sa.Column('meeting_result', sa.Enum('EXCELLENT', 'GOOD', 'BADLY', 'DISGUSTING', name='resultmeetingenum'), nullable=False),
-    sa.Column('participant_engagement', sa.Boolean(), nullable=False),
-    sa.Column('problem_solution', sa.Boolean(), nullable=False),
-    sa.Column('meeting_feedback', sa.Text(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['meeting_id'], ['meeting.id'], ),
-    sa.ForeignKeyConstraint(['owner_id'], ['usertabit.id'], ),
-    sa.PrimaryKeyConstraint('id', 'meeting_id'),
-    sa.UniqueConstraint('id')
-    )
-    op.create_table('votingfeed',
-    sa.Column('message_id', sa.Integer(), nullable=False),
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=30), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['message_id'], ['messagefeed.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id'),
-    sa.UniqueConstraint('name')
-    )
-    op.create_table('votingbyuser',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('user_id', fastapi_users_db_sqlalchemy.generics.GUID(), nullable=True),
-    sa.Column('voting_id', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['usertabit.id'], ),
-    sa.ForeignKeyConstraint(['voting_id'], ['votingfeed.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('votingbyuser')
-    op.drop_table('votingfeed')
-    op.drop_table('resultmeeting')
-    op.drop_table('filetask')
-    op.drop_table('filemessage')
-    op.drop_table('filemeeting')
-    op.drop_table('commentfeed')
-    op.drop_table('associationusertask')
-    op.drop_table('associationusermeeting')
-    op.drop_table('task')
-    op.drop_table('messagefeed')
-    op.drop_table('meeting')
-    op.drop_table('fileproblem')
-    op.drop_table('associationuserproblem')
-    op.drop_table('problem')
-    # ### end Alembic commands ###
