--- alembic/versions/07_bugfix_department_taguser.py
@@ -1,32 +1,0 @@
-"""Bugfix: Department, TagUser
-
-Revision ID: 07
-Revises: 06
-Create Date: 2025-01-26 02:26:48.089230
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '07'
-down_revision: Union[str, None] = '06'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('taguser', sa.Column('company_id', sa.Integer(), nullable=True))
-    op.create_foreign_key(None, 'taguser', 'company', ['company_id'], ['id'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'taguser', type_='foreignkey')
-    op.drop_column('taguser', 'company_id')
-    # ### end Alembic commands ###
