--- alembic/versions/08_add_landing_page_and_field_slug.py
@@ -1,54 +1,0 @@
-"""add_landing_page_and_field_slug
-
-Revision ID: 08
-Revises: 07
-Create Date: 2025-01-27 23:26:05.725076
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '08'
-down_revision: Union[str, None] = '07'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('landingpage',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('phone_number_1', sa.String(), nullable=True),
-    sa.Column('phone_number_2', sa.String(), nullable=True),
-    sa.Column('phone_number_3', sa.String(), nullable=True),
-    sa.Column('address', sa.String(), nullable=True),
-    sa.Column('email', sa.String(), nullable=True),
-    sa.Column('whatsapp', sa.String(), nullable=True),
-    sa.Column('telegram', sa.String(), nullable=True),
-    sa.Column('vk', sa.String(), nullable=True),
-    sa.Column('price_1', sa.String(), nullable=True),
-    sa.Column('price_2', sa.String(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('id')
-    )
-    op.add_column('company', sa.Column('slug', sa.String(length=25), nullable=False))
-    op.create_unique_constraint(None, 'company', ['slug'])
-    op.add_column('department', sa.Column('slug', sa.String(length=25), nullable=False))
-    op.create_unique_constraint(None, 'department', ['slug'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'department', type_='unique')
-    op.drop_column('department', 'slug')
-    op.drop_constraint(None, 'company', type_='unique')
-    op.drop_column('company', 'slug')
-    op.drop_table('landingpage')
-    # ### end Alembic commands ###
