--- alembic/versions/09_bugfix_null_field.py
@@ -1,180 +1,0 @@
-"""bugfix_null_field
-
-Revision ID: 09
-Revises: 08
-Create Date: 2025-01-31 03:50:39.417138
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '09'
-down_revision: Union[str, None] = '08'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('commentfeed', 'owner_id',
-               existing_type=sa.UUID(),
-               nullable=False)
-    op.alter_column('company', 'logo',
-               existing_type=sa.VARCHAR(length=2048),
-               nullable=True)
-    op.alter_column('filemeeting', 'file_path',
-               existing_type=sa.VARCHAR(length=2048),
-               nullable=True)
-    op.alter_column('filemessage', 'file_path',
-               existing_type=sa.VARCHAR(length=2048),
-               nullable=True)
-    op.alter_column('fileproblem', 'file_path',
-               existing_type=sa.VARCHAR(length=2048),
-               nullable=True)
-    op.alter_column('filetask', 'file_path',
-               existing_type=sa.VARCHAR(length=2048),
-               nullable=True)
-    op.create_unique_constraint(None, 'landingpage', ['id'])
-    op.alter_column('meeting', 'owner_id',
-               existing_type=sa.UUID(),
-               nullable=False)
-    op.alter_column('messagefeed', 'owner_id',
-               existing_type=sa.UUID(),
-               nullable=False)
-    op.alter_column('problem', 'owner_id',
-               existing_type=sa.UUID(),
-               nullable=False)
-    op.alter_column('resultmeeting', 'owner_id',
-               existing_type=sa.UUID(),
-               nullable=False)
-    op.alter_column('tabitadminuser', 'patronymic',
-               existing_type=sa.VARCHAR(length=100),
-               nullable=True)
-    op.alter_column('tabitadminuser', 'phone_number',
-               existing_type=sa.VARCHAR(),
-               nullable=True)
-    op.alter_column('taguser', 'company_id',
-               existing_type=sa.INTEGER(),
-               nullable=False)
-    op.alter_column('task', 'owner_id',
-               existing_type=sa.UUID(),
-               nullable=False)
-    op.alter_column('usertabit', 'birthday',
-               existing_type=sa.DATE(),
-               nullable=True)
-    op.alter_column('usertabit', 'telegram_username',
-               existing_type=sa.VARCHAR(length=100),
-               nullable=True)
-    op.alter_column('usertabit', 'start_date_employment',
-               existing_type=sa.DATE(),
-               nullable=True)
-    op.alter_column('usertabit', 'end_date_employment',
-               existing_type=sa.DATE(),
-               nullable=True)
-    op.alter_column('usertabit', 'current_department_id',
-               existing_type=sa.INTEGER(),
-               nullable=True)
-    op.alter_column('usertabit', 'department_transition_date',
-               existing_type=sa.DATE(),
-               nullable=True)
-    op.alter_column('usertabit', 'employee_position',
-               existing_type=sa.VARCHAR(),
-               nullable=True)
-    op.alter_column('usertabit', 'avatar_link',
-               existing_type=sa.VARCHAR(length=2048),
-               nullable=True)
-    op.alter_column('usertabit', 'patronymic',
-               existing_type=sa.VARCHAR(length=100),
-               nullable=True)
-    op.alter_column('usertabit', 'phone_number',
-               existing_type=sa.VARCHAR(),
-               nullable=True)
-    op.alter_column('votingbyuser', 'user_id',
-               existing_type=sa.UUID(),
-               nullable=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('votingbyuser', 'user_id',
-               existing_type=sa.UUID(),
-               nullable=True)
-    op.alter_column('usertabit', 'phone_number',
-               existing_type=sa.VARCHAR(),
-               nullable=False)
-    op.alter_column('usertabit', 'patronymic',
-               existing_type=sa.VARCHAR(length=100),
-               nullable=False)
-    op.alter_column('usertabit', 'avatar_link',
-               existing_type=sa.VARCHAR(length=2048),
-               nullable=False)
-    op.alter_column('usertabit', 'employee_position',
-               existing_type=sa.VARCHAR(),
-               nullable=False)
-    op.alter_column('usertabit', 'department_transition_date',
-               existing_type=sa.DATE(),
-               nullable=False)
-    op.alter_column('usertabit', 'current_department_id',
-               existing_type=sa.INTEGER(),
-               nullable=False)
-    op.alter_column('usertabit', 'end_date_employment',
-               existing_type=sa.DATE(),
-               nullable=False)
-    op.alter_column('usertabit', 'start_date_employment',
-               existing_type=sa.DATE(),
-               nullable=False)
-    op.alter_column('usertabit', 'telegram_username',
-               existing_type=sa.VARCHAR(length=100),
-               nullable=False)
-    op.alter_column('usertabit', 'birthday',
-               existing_type=sa.DATE(),
-               nullable=False)
-    op.alter_column('task', 'owner_id',
-               existing_type=sa.UUID(),
-               nullable=True)
-    op.alter_column('taguser', 'company_id',
-               existing_type=sa.INTEGER(),
-               nullable=True)
-    op.alter_column('tabitadminuser', 'phone_number',
-               existing_type=sa.VARCHAR(),
-               nullable=False)
-    op.alter_column('tabitadminuser', 'patronymic',
-               existing_type=sa.VARCHAR(length=100),
-               nullable=False)
-    op.alter_column('resultmeeting', 'owner_id',
-               existing_type=sa.UUID(),
-               nullable=True)
-    op.alter_column('problem', 'owner_id',
-               existing_type=sa.UUID(),
-               nullable=True)
-    op.alter_column('messagefeed', 'owner_id',
-               existing_type=sa.UUID(),
-               nullable=True)
-    op.alter_column('meeting', 'owner_id',
-               existing_type=sa.UUID(),
-               nullable=True)
-    op.drop_constraint(None, 'landingpage', type_='unique')
-    op.alter_column('filetask', 'file_path',
-               existing_type=sa.VARCHAR(length=2048),
-               nullable=False)
-    op.alter_column('fileproblem', 'file_path',
-               existing_type=sa.VARCHAR(length=2048),
-               nullable=False)
-    op.alter_column('filemessage', 'file_path',
-               existing_type=sa.VARCHAR(length=2048),
-               nullable=False)
-    op.alter_column('filemeeting', 'file_path',
-               existing_type=sa.VARCHAR(length=2048),
-               nullable=False)
-    op.alter_column('company', 'logo',
-               existing_type=sa.VARCHAR(length=2048),
-               nullable=False)
-    op.alter_column('commentfeed', 'owner_id',
-               existing_type=sa.UUID(),
-               nullable=True)
-    # ### end Alembic commands ###
