--- alembic/versions/10_redesgn_relation_department.py
@@ -1,36 +1,0 @@
-"""redesgn_relation_department
-
-Revision ID: 10
-Revises: 09
-Create Date: 2025-02-02 08:56:46.910550
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '10'
-down_revision: Union[str, None] = '09'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('department_supervisor_id_fkey', 'department', type_='foreignkey')
-    op.drop_column('department', 'supervisor_id')
-    op.add_column('usertabit', sa.Column('supervisor', sa.Boolean(), nullable=True))
-    op.create_unique_constraint('unique_supervisor', 'usertabit', ['supervisor', 'current_department_id'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('unique_supervisor', 'usertabit', type_='unique')
-    op.drop_column('usertabit', 'supervisor')
-    op.add_column('department', sa.Column('supervisor_id', sa.UUID(), autoincrement=False, nullable=True))
-    op.create_foreign_key('department_supervisor_id_fkey', 'department', 'usertabit', ['supervisor_id'], ['id'])
-    # ### end Alembic commands ###
