--- alembic/versions/11_datetime_for_company.py
@@ -1,44 +1,0 @@
-"""DateTime for Company
-
-Revision ID: 11
-Revises: 10
-Create Date: 2025-02-04 00:19:00.646502
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '11'
-down_revision: Union[str, None] = '10'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('company', 'start_license_time',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=True)
-    op.alter_column('company', 'end_license_time',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('company', 'end_license_time',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True)
-    op.alter_column('company', 'start_license_time',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True)
-    # ### end Alembic commands ###
