--- alembic/versions/12_datetime_for_company.py
@@ -1,488 +1,0 @@
-"""DateTime for Company
-
-Revision ID: 12
-Revises: 11
-Create Date: 2025-02-04 00:45:56.176655
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '12'
-down_revision: Union[str, None] = '11'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('associationusermeeting', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationusermeeting', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationuserproblem', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationuserproblem', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationusertags', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationusertags', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationusertask', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationusertask', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('commentfeed', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('commentfeed', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('company', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('company', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('department', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('department', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('filemeeting', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('filemeeting', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('filemessage', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('filemessage', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('fileproblem', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('fileproblem', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('filetask', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('filetask', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('landingpage', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('landingpage', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('licensetype', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('licensetype', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('meeting', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('meeting', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('messagefeed', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('messagefeed', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('problem', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('problem', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('resultmeeting', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('resultmeeting', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('tabitadminuser', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('tabitadminuser', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('taguser', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('taguser', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('task', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('task', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('usertabit', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('usertabit', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('votingbyuser', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('votingbyuser', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('votingfeed', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('votingfeed', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=postgresql.TIMESTAMP(timezone=True),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('votingfeed', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('votingfeed', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('votingbyuser', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('votingbyuser', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('usertabit', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('usertabit', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('task', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('task', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('taguser', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('taguser', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('tabitadminuser', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('tabitadminuser', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('resultmeeting', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('resultmeeting', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('problem', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('problem', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('messagefeed', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('messagefeed', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('meeting', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('meeting', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('licensetype', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('licensetype', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('landingpage', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('landingpage', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('filetask', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('filetask', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('fileproblem', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('fileproblem', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('filemessage', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('filemessage', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('filemeeting', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('filemeeting', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('department', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('department', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('company', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('company', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('commentfeed', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('commentfeed', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationusertask', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationusertask', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationusertags', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationusertags', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationuserproblem', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationuserproblem', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationusermeeting', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    op.alter_column('associationusermeeting', 'created_at',
-               existing_type=postgresql.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False,
-               existing_server_default=sa.text('now()'))
-    # ### end Alembic commands ###
