+++ src/api/v1/endpoints/tabit_management_companies.py
@@ -2,20 +2,20 @@
 Эндпоинты управления компаниями (Tabit Management - Companies).
 """
 
+from http import HTTPStatus
 from typing import List
 
 from fastapi import APIRouter, Depends
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from src.database.db_depends import get_async_session
+from src.api.v1.validator import validator_check_object_exists
 from src.companies.crud import company_crud
 from src.companies.schemas import (
     CompanyCreateSchema,
-    CompanyResponseForUserSchema,
     CompanyResponseSchema,
     CompanyUpdateSchema,
-    CompanyUpdateForUserSchema,
 )
+from src.database.db_depends import get_async_session
 
 router = APIRouter()
 
@@ -23,6 +23,7 @@ router = APIRouter()
 @router.get(
     '/',
     response_model=List[CompanyResponseSchema],
+    # TODO: Зависимость супер_админ_табит.
     summary='Получить список всех компаний',
 )
 async def get_companies(
@@ -37,6 +38,8 @@ async def get_companies(
 @router.post(
     '/',
     response_model=CompanyResponseSchema,
+    # TODO: Зависимость супер_админ_табит.
+    status_code=HTTPStatus.CREATED,
     summary='Создать новую компанию',
 )
 async def create_company(
@@ -46,28 +49,33 @@ async def create_company(
     """
     Создает новую компанию. Доступно только админам сервиса.
     """
-    new_company = await company_crud.create(session, company)
-    return new_company
+    return await company_crud.create(session, company)
 
 
 @router.patch(
     '/{company_slug}/',
+    response_model=CompanyResponseSchema,
+    # TODO: Зависимость супер_админ_табит.
     summary='Обновить данные компании',
 )
 async def update_company(
     company_slug: str,
+    object_in: CompanyUpdateSchema,
     session: AsyncSession = Depends(get_async_session),
 ):
     """
-    Обновляет данные компании по её slug.
+    Обновляет данные компании по её slug. Доступно только админам сервиса.
     :param company_slug: Уникальный идентификатор компании (slug).
     """
-    # TODO: Реализовать обновление данных компании
-    return {'message': f'Обновление компании {company_slug} временно недоступно'}
+    company = await validator_check_object_exists(
+        session, company_crud, object_slug=company_slug
+    )
+    return await company_crud.update(session, company, object_in)
 
 
 @router.delete(
     '/{company_slug}/',
+    response_model=CompanyResponseSchema,
     summary='Удалить компанию',
 )
 async def delete_company(
@@ -75,8 +83,10 @@ async def delete_company(
     session: AsyncSession = Depends(get_async_session),
 ):
     """
-    Удаляет компанию по её slug.
+    Удаляет компанию по её slug. Доступно только админам сервиса.
     :param company_slug: Уникальный идентификатор компании (slug).
     """
-    # TODO: Реализовать удаление компании
-    return {'message': f'Удаление компании {company_slug} временно недоступно'}
+    company = await validator_check_object_exists(
+        session, company_crud, object_slug=company_slug
+    )
+    return await company_crud.remove(session, company)
